% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAlleles}
\alias{GetAlleles}
\title{Compute .. allels}
\usage{
GetAlleles(Both, Allele, n_sample)
}
\arguments{
\item{Both}{A character vector vector}

\item{Allele}{An integer matrix}

\item{n_sample}{Number of elements}
}
\description{
Function GetAlleles description
}
