% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histSampCov.R
\name{histSampCov}
\alias{histSampCov}
\title{histSampCov function}
\usage{
histSampCov(coverage, loci = c(1:16569))
}
\arguments{
\item{coverage}{a numeric matrix (16569 x N) containing the subject ID as the row names, and the reads
coverage of the 16569 mitochondrial DNA loci for each subject.}

\item{loci}{one of: 1. a vector(default is c(1:16569)) of mitochondrial DNA loci to specify which loci
should be used to identify the variations and annotate, 2. a character string for the regions
(e.g. “coding”, “tRNA”, “RNR1”, "RNR2", …).}
}
\description{
Produce a Histogram of Mean Coverage Across Subjects
}
\examples{


#histSampCov (coverage)
#histSampCov (coverage, loci="tRNA")

}
