% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAnalysis.R
\name{mtAnalysis}
\alias{mtAnalysis}
\title{mtAnalysis function}
\usage{
mtAnalysis(
  aaf,
  family,
  Methods,
  pheno,
  trait,
  covars,
  G_coding,
  region,
  aaf_cat_SNPInfo = aaf_cat_SNPInfo,
  kins = NULL
)
}
\arguments{
\item{aaf}{test1}

\item{pheno}{test4}

\item{covars}{test2}

\item{G_coding}{test5}

\item{aaf_cat_SNPInfo}{test3}

\item{kins}{}
}
\value{

}
\description{
mtAnalysis function
}
