% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAnno.R
\name{mtAnno}
\alias{mtAnno}
\title{mtAnno function}
\usage{
mtAnno(
  anno,
  annot.select = c("Pos", "ref", "Gene", "TypeMutation", "MissensMutation",
    "CodonPosition", "ProteinDomain", "dbSNP_150_id", "PolyPhen2", "PolyPhen2_score",
    "SIFT", "SIFT_score", "CADD", "CADD_score", "CADD_phred_score"),
  path = "./",
  study = "Study"
)
}
\arguments{
\item{anno}{data frame provided by the user. It contains alleles need to be annotated.
It has two columns: loci positions ("pos") and "alleles" to be annotated.}

\item{annot.select}{types of annotation scores based on user's choice}

\item{path}{the path of the output annotation file. If not provided, the annotation file will
output to the current working directory}

\item{study}{A string of study names. Default is Study.}
}
\value{
Annotation of alleles stored as a .csv file to the path provided by users
}
\description{
Annotate mtDNA alternative alleles
}
\examples{

 # mtAnno(anno)
}
