% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtdna_test.R
\name{mtdna_test}
\alias{mtdna_test}
\title{mtdna test function}
\usage{
mtdna_test(coverage_file, allele_file, freq_file, path)
}
\arguments{
\item{coverage_file}{Coverage File}

\item{allele_file}{Allele File}

\item{freq_file}{Frequency File}

\item{path}{Path}
}
\value{
character
}
\description{
Test function for the package
}
\examples{
coverage_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/coverage/coverage.csv"
allele_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/allele/allele.csv"
freq_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/freq/freq.csv"
path<-"/rprojectnb2/mtdna-alcohol/Sun_Xianbang/Annotation/Output/"
#mtdna_test (coverage_file, allele_file, freq_file)
}
