% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtPAA.R
\name{mtPAA}
\alias{mtPAA}
\title{mtPAA function}
\usage{
mtPAA(allele, freq)
}
\arguments{
\item{allele}{Allele dataset, numeric matrix}

\item{freq}{Frequency dataset, numeric matrix}
}
\value{
PAA, numeric matrix with dimensions of n subjects x 16569 SNPs
}
\description{
Derivation of alternative allele fraction
}
\examples{

#Read input data
allele_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/allele/allele.csv"
freq_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/freq/freq.csv"

#head <- scan( file = allele_file, sep = ",", character(), nlines = 1, quiet = TRUE)
#allele   <- matrix( scan( file = allele_file,
#                          sep = ",", character() ),
#                    ncol = length( head ), byrow = TRUE)
#freq     <- matrix( scan( file = freq_file,
#                          sep = ",", character() ),
#                    ncol = length( head ), byrow = TRUE)

#mtPAA (coverage, allele, freq)
}
