% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtReadVCF.R
\name{mtReadVCF}
\alias{mtReadVCF}
\title{mtReadVCF function}
\usage{
mtReadVCF(VCF_file, n_sample)
}
\arguments{
\item{VCF_file}{name of VCF file to import}

\item{n_sample}{number of subjects}
}
\value{

}
\description{
Convert VCF to allele, freq and coverage datasets
}
\examples{

}
