% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCover.R
\name{plotCover}
\alias{plotCover}
\title{plotCover function}
\usage{
plotCover(
  x,
  loci = c(1:.mtLength),
  col = "blue",
  pch = ".",
  cex = 0.2,
  xlab = "mtloci",
  ylab = "",
  main = "Mean Coverage across all mtDNA loci",
  ...
)
}
\arguments{
\item{x}{a numeric matrix (16569 x N). Rows correspond to loci and columns correspond to subjects. It
contains the subject ID as the column names, and the reads coverage of the 16569 mtDNA loci for each
subject.}

\item{loci}{one of the following to specify mtDNA loci: 1. a numeric vector (default is c(1:16569)) of
mitochondrial DNA loci, 2. a character string for the regions(e.g. "coding", "tRNA", "Dloop", …).}

\item{col}{color to be used in the plot.}

\item{pch}{plotting ‘character’, i.e., symbol to use.}

\item{cex}{expansion factor for symbols used in the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{main}{a title for the plot.}

\item{...}{arguments to be passed to plot function.}
}
\description{
Produce a Diagnostic Scatter Plot to Visualize Mean Coverage Across Subjects for Each Locus
}
\examples{


#plotCover (coverage)
#plotCover (coverage, loci="tRNA")

}
