% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLociMeanCoverage.R
\name{plotLociMeanCoverage}
\alias{plotLociMeanCoverage}
\title{plotLociMeanCoverage function}
\usage{
plotLociMeanCoverage(
  x,
  loci = c(1:16569),
  col = "blue",
  pch = ".",
  cex = 0.2,
  xlab = "mtloci",
  ylab = "",
  main = "Mean Coverage across all mtDNA loci",
  ...
)
}
\arguments{
\item{x}{a numeric matrix (16569 x N) containing the subject ID as the column names,
and the reads coverage of the 16569 mitochondrial DNA loci for each subject.}

\item{loci}{one of: 1. a vector(default is c(1:16569)) of mitochondrial DNA loci to specify which loci
should be used to identify the variations and annotate, 2. a character string for the regions
(e.g. “coding”, “tRNA”, “RNR1”, "RNR2", …).}

\item{col}{color to be used in the plot.}

\item{pch}{plotting ‘character’, i.e., symbol to use.}

\item{cex}{expansion factor for symbols used in the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{main}{a title for the plot.}

\item{...}{arguments to be passed to plot function.}
}
\description{
Produce a scatter plot of Mean Coverage Across Loci
}
\examples{


#plotLociMeanCoverage (coverage)
#plotLociMeanCoverage (coverage, loci="tRNA")

}
