% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histSampCov.R
\name{histSampCov}
\alias{histSampCov}
\title{histSampCov function}
\usage{
histSampCov(coverage, loci = c(1:.mtLength))
}
\arguments{
\item{coverage}{a numeric matrix (16569 x N). Rows correspond to loci and columns correspond to subjects.
It contains the subject ID as the column names, and the reads coverage of the 16569 mtDNA loci for each
subject.}

\item{loci}{one of the following to specify mtDNA loci: 1. a numeric vector (default is c(1:16569)) of
mitochondrial DNA loci, 2. a character string for the regions(e.g. "coding", "tRNA", "Dloop", …).}
}
\description{
Produce a Histogram of Mean Coverage Across mtDNA Loci for Subjects
}
\examples{


#histSampCov (coverage)
#histSampCov (coverage, loci="tRNA")

}
