% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAAF.R
\name{mtAAF}
\alias{mtAAF}
\title{mtAAF function}
\usage{
mtAAF(allele, freq)
}
\arguments{
\item{allele}{data frame (N x 16569) provided by the user. This data frame contains
N subjects with mtDNA sequencing data of 16569 loci. The data frame contains subject ID as
the row names, and the allele calls of all mtDNA loci for each subject as the columns.
“/” is used to delimited different allele calls in a locus.}

\item{freq}{data frame (N x 16569) provided by the user. This data frame contains
N subjects with mtDNA sequencing data of 16569 loci. The data frame contains subject ID as
the row names, and the allele fractions of the called alleles for each subject as the columns.
“/” is used to delimited the allele fractions.}
}
\value{
AAF, numeric matrix (N x 16569). It contains subject ID as the row names,
and the AAF of all 16569 mtDNA loci for each subject. It will also generate a scatter plot
of AAF based on user’s choice.
}
\description{
Derivation of alternative allele fraction
}
\examples{

#Read input data
allele_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/allele/allele.csv"
freq_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/freq/freq.csv"

#head <- scan( file = allele_file, sep = ",", character(), nlines = 1, quiet = TRUE)
#allele   <- matrix( scan( file = allele_file,
#                          sep = ",", character() ),
#                    ncol = length( head ), byrow = TRUE)
#freq     <- matrix( scan( file = freq_file,
#                          sep = ",", character() ),
#                    ncol = length( head ), byrow = TRUE)

#mtPAA ( allele, freq)
}
