% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAAF.R
\name{mtAAF}
\alias{mtAAF}
\title{mtAAF function}
\usage{
mtAAF(allele, freq, method = "maxAA")
}
\arguments{
\item{allele}{a character matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
This matrix contains the alleles of each subject at each locus.
The matrix must contain subject ID as the column names.
"/" is used to delimited different allele calls in a locus.}

\item{freq}{a character matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
This matrix contains the allele fractions of the corresponding allele matrix.
The matrix must contain subject ID as the column names.
"/" is used to delimited the allele fractions.}

\item{method}{the method to compute AAF for the case of multiple alternative
alleles. The default method "maxAA" computes AAF as the maximum of
frequencies of corresponding alternative alleles; and the alternative method
"allAA" computes AAF as 1 minus the frequency of reference allele}
}
\value{
AAF, a numeric matrix (16569 x N).
Rows correspond to loci and columns correspond to subjects.
It contains subject ID as the column names, and the AAFs of all 16569 mtDNA
loci for each subject.
}
\description{
Derivation of alternative allele fraction (AAF).
}
\examples{

\dontrun{
## Read input data
allele_file <- "allele.csv"
freq_file   <- "freq.csv"

allele <- as.matrix( read.csv(file=allele_file, sep=",", header=FALSE))
freq   <- as.matrix( read.csv(file=freq_file, sep=",", header=FALSE))

aaf=mtAAF(allele, freq)
}

}
