% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAnalysis.R
\name{mtAnalysis}
\alias{mtAnalysis}
\title{Title}
\usage{
mtAnalysis(
  aaf,
  family,
  Methods,
  pheno,
  trait,
  covars,
  G_coding,
  region,
  aaf_cat_SNPInfo = aaf_cat_SNPInfo,
  kins = NULL
)
}
\arguments{
\item{kins}{}
}
\value{

}
\description{
Title
}
