% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtAnno.R
\name{mtAnno}
\alias{mtAnno}
\title{mtAnno function}
\usage{
mtAnno(
  anno,
  annot.select = c("Pos", "ref", "Gene", "TypeMutation", "MissensMutation",
    "CodonPosition", "ProteinDomain", "dbSNP_150_id", "PolyPhen2", "PolyPhen2_score",
    "SIFT", "SIFT_score", "CADD", "CADD_score", "CADD_phred_score"),
  path = "./",
  study = "Study"
)
}
\arguments{
\item{anno}{a data frame provided by the user. It has two columns:
mtDNA loci positions ("pos") and
alternative alleles ("alleles") to be annotated.}

\item{annot.select}{A character vector of variation position, alternative
allele, corresponding gene and types of annotation scores to output based on
user's choice. The available choices are "Pos", "ref", "Gene", "TypeMutation",
"MissensMutation", "CodonPosition", "ProteinDomain", "mFOLD_dG",
"mFOLD_Initial", "mFOLD_rCRS.DG", "mFOLD_rCRS.Initial",
"mFOLD_AnticodonAminoAcidChange", "mFOLD_Location", "PolyPhen2",
"PolyPhen2_score", "SIFT", "SIFT_score", "PROVEAN", "PROVEAN_score",
"MutationAssessor", "MutationAssessor_score", "CADD", "CADD_score",
"CADD_phred_score", "PANTHER", "PANTHER_score", "PhD_SNP", "PhD_SNP_score",
"SNAP", "SNAP_score", "MutationTaster", "MutationTaster_score", "dbSNP_150_id",
"APOGEE_boost_consensus", "APOGEE_boost_mean_prob", "APOGEE_boost_mean"}

\item{path}{the path to the directory of the output of annotation file.
If not provided, the annotation
file will output to the current working directory.}

\item{study}{A string of study names. Default is "Study".}
}
\value{
A .csv file containing annotated alternative alleles with
corresponding mtDNA loci positions.
}
\description{
Annotate Given Alternative Alleles of mtDNA Loci
}
\examples{

\dontrun{
## Read input data
allele_file <- "allele.csv"
freq_file   <- "freq.csv"

allele <- as.matrix( read.csv(file = allele_file, sep = ",", header=FALSE) )
freq   <- as.matrix( read.csv(file = freq_file, sep = ",", header=FALSE) )

aaf =  mtAAF ( allele, freq)

## Annotate Given Alternative Alleles of mtDNA Loci
mtAnno(aaf)

##
mtAnno(aaf, annot.select=c("ref","Gene") )
}
}
