% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtdna_test.R
\name{mtdna_test}
\alias{mtdna_test}
\title{mtdna test function}
\usage{
mtdna_test(head_file, coverage_file, allele_file, freq_file)
}
\arguments{
\item{head_file}{Head File}

\item{coverage_file}{Coverage File}

\item{allele_file}{Allele File}

\item{freq_file}{Frequency File}
}
\value{
character
}
\description{
Test function for the package
}
\examples{
head_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/allele/allele.csv"
coverage_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/coverage/coverage.csv"
allele_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/allele/allele.csv"
freq_file <- "/restricted/projectnb/mtdna-alcohol/Sun_Xianbang/ARIC/freq/freq.csv"
mtdna_test (head_file, coverage_file, allele_file, freq_file)
}
