% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtdna_test.R
\name{mtdna_test}
\alias{mtdna_test}
\title{mtdna test function}
\usage{
mtdna_test(coverage_file, allele_file, freq_file, path)
}
\arguments{
\item{coverage_file}{Coverage File}

\item{allele_file}{Allele File}

\item{freq_file}{Frequency File}

\item{path}{Path}
}
\value{
character
}
\description{
Test function for the package
}
\examples{

\dontrun{
#Read input data
coverage_file <- "coverage.csv"
allele_file <- "allele.csv"
freq_file   <- "freq.csv"

coverage <- as.matrix( read.csv(file=coverage_file, sep=",", header=FALSE))
allele <- as.matrix( read.csv(file=allele_file, sep=",", header=FALSE))
freq   <- as.matrix( read.csv(file=freq_file, sep=",", header=FALSE))

mtdna_test(coverage_file=coverage, allele_file=allele, freq_file=freq)
}

}
