% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtReadVCF_mtc.R
\name{mtReadVCF_mtc}
\alias{mtReadVCF_mtc}
\title{mtReadVCF_mtc function}
\usage{
mtReadVCF_mtc(VCF_file, n_sample)
}
\arguments{
\item{VCF_file}{a string of path and name of VCF file to import}

\item{n_sample}{number of subjects}
}
\value{
A list contains the three matrices of allele, freq and coverage
with dimensions of 16569 x N. Rows correspond to loci and columns correspond
to subjects. It contains subject ID as the column names.
}
\description{
Extract allele, freq and coverage datasets from the VCF file containing
results called by mitoCaller. The datasets are the input of mtAAF and
mtSummary functions.
}
\examples{

}
