% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtSummary.R
\name{mtSummary}
\alias{mtSummary}
\title{mtSummary function}
\usage{
mtSummary(
  aaf,
  allele,
  freq,
  coverage,
  coverage.qc = 250,
  thre.lower = 0.03,
  thre.upper = 0.97,
  loci = c(1:.mtLength),
  type = "both",
  coverSummary = T,
  varHist = T,
  annot.select = c("Pos", "ref", "Gene", "TypeMutation", "MissensMutation",
    "CodonPosition", "ProteinDomain", "dbSNP_150_id", "PolyPhen2", "PolyPhen2_score",
    "SIFT", "SIFT_score", "CADD", "CADD_score", "CADD_phred_score"),
  path = "./",
  study = "Study",
  anno = T
)
}
\arguments{
\item{aaf}{a numeric matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
It contains subject ID as the column names,
and the AAFs of all 16569 mtDNA loci for each subject.
It is generated from mtAAF function.}

\item{allele}{a character matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
This matrix contains the alleles of each subject at each locus.
The matrix must contain subject ID as the column names.
"/" is used to delimited different allele calls in a locus.}

\item{freq}{a character matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
This matrix contains the allele fractions of the corresponding allele matrix.
The matrix must contain subject ID as the column names.
"/" is used to delimited the allele fractions.}

\item{coverage}{a numeric matrix (16569 x N) provided by the user.
Rows correspond to loci and columns correspond to subjects.
This matrix contains the reads coverage of the 16569 mtDNA loci for each
subject. The matrix must contain the subject ID as the column names.}

\item{coverage.qc}{a number(default is 250) of threshold for the coverage.
If the coverage<coverage.qc, the allele call at that locus of the subject
will not be used.}

\item{thre.lower}{a number(default is 0.03) of lower bound of the threshold
defining heteroplasmic
and homoplasmic variations}

\item{thre.upper}{a number(default is 0.97) of upper bound of the threshold
defining heteroplasmic
and homoplasmic variations}

\item{loci}{one of:
1. a vector(default is c(1:16569)) of mitochondrial DNA  loci to specify
which loci should be used to identify the variations and annotate,
2. a character string for the regions
(e.g. "coding" , "tRNA", "RNR1" , "RNR2",...)}

\item{type}{a character of indicator choosing to output annotation
to all variations,
heteroplasmic variations, or homoplasmic variations. “both” returns
annotation to all
variations (default), "heter" returns annotation to heteroplasmic
variations and "homo"
returns annotation to homoplasmic variations.}

\item{coverSummary}{logical(default is True). A user can specify to output
summary of mean coverage at each mtDNA loci (across all participants) and
for each individual (across all mtDNA loci).}

\item{varHist}{logical(default is True). A user can specify to output
histograms to visualize the
heteroplasmic and homoplasmic burden across participants and mtDNA loci.}

\item{annot.select}{types of annotation scores based on user's choice}

\item{path}{the path of the output annotation file. If not provided,
the annotation file will
output to the current working directory}

\item{study}{A string of study names. Default is Study.}

\item{anno}{A logical value (default is False) indicating whether output the
annotation results in a .csv file}
}
\value{
1. Summary frequency: descriptive statistics of sequencing coverage
across individual
and across mtDNA loci; the total number of mtDNA loci with variations and
the number of
heteroplasmic/homoplasmic mtDNA loci in the study sample;
the min/Q1/mean/median/Q3/largest number
of homoplasmy/heteroplasmy carried by an individual.
2. Plots: a scatter plot of the median coverage
across mtDNA loci; histograms to visualize the heteroplasmic and
homoplasmic burden across
participants and mtDNA loci based on user’s choice.
3.	Summary annotation for all of the
heteroplasmic and/or homoplasmic variations observed in
the study data based on user's choice.
}
\description{
Identification of mtDNA variations, output summary statistics,
annotation of heteroplasmic and/or homoplasmic variations.
}
\examples{

\dontrun{
## Read input data
allele_file <- "allele.csv"
freq_file   <- "freq.csv"

allele <- as.matrix( read.csv(file = allele_file, sep = ",", header=FALSE) )
freq   <- as.matrix( read.csv(file = freq_file, sep = ",", header=FALSE) )

aaf =  mtAAF ( allele, freq)

## Summary for the aaf object
mtSummary(aaf, allele, freq, coverage)
}
}
