% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mtDNAaaf.R
\name{plot.mtDNAaaf}
\alias{plot.mtDNAaaf}
\title{plot.mtDNAaaf function}
\usage{
\method{plot}{mtDNAaaf}(x, col = "blue", pch = ".", cex = 0.2, xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{numeric matrix (16569 x N ). It contains subject ID as the column names,
and the AAF of all 16569 mtDNA loci for each subject.}

\item{col}{color to be used in the plot.}

\item{pch}{plotting ‘character’, i.e., symbol to use.}

\item{cex}{expansion factor for symbols used in the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{...}{arguments to be passed to plot function.}
}
\description{
Produce a scatter plot of a mtDNAaaf object
}
\examples{


#plot (aaf)

}
